Extract TaskPriority to its own header file
